requirements.txt
keras==2.12.0
numpy==1.23.5
Pillow==11.2.1
scikit_learn==1.6.1
tensorflow==2.12.0
